+++ CHANGELOG.md
@@ -6,14 +6,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Initialize
+
+    build (1)
 - Implemented - Deposit
 - Implemented - Withdrawal
 - Implemented - Transfer
 
+    build (2)
 - Changed logic related to deposit flow
 - Fixed user list item if user doesn't have username
+
+    build (3)
 - Removed switch network button on deposit details screen
 - Hide zero balanced coins on TransferAssetScreen
 - Added validation to transfer's input
 - Fixed crash on transfer
 - Added Network logger
+
+    build(4)
+- Added refresh control